Moved newsletters to _posts
Imported the newsletters as jekyll posts, instead of static pages, with
the category of 'newsletter'.  The jekyll permalink metadata is used to
generate the same urls as the old site.  The files were renamed to the
jekyll mandated year-mon-day-title format for posts.  The newsletter
index is automatically generated now, and the extra newsletter.yml file
is not needed, so was deleted.